Fix: Remove invalid toLong() on currentSong and ensure currentPosition is Long

+++ app/src/main/java/com/subho/aurabeat/MainActivity.kt
@@ -51,7 +51,7 @@ class MainActivity : ComponentActivity() {
 
                 MusicScreen(
                     songs = songs,
-                    currentSong = currentSong.toLong(),
+                    currentSong = currentSong,
                     isPlaying = isPlaying,
                     currentPosition = currentPosition.toLong(),
                     duration = duration.toFloat(),
